# BackEnd/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PositiveQuote",
    "version": "v1"
  },
  "paths": {
    "/api/Quotes": {
      "get": {
        "tags": [
          "Quotes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Quote"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quotes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        }
      }
    },
    "/api/Quotes/{id}": {
      "get": {
        "tags": [
          "Quotes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Quotes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Quote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Quotes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Quote": {
      "required": [
        "quote_Content"
      ],
      "type": "object",
      "properties": {
        "quote_ID": {
          "format": "int32",
          "type": "integer"
        },
        "quote_Content": {
          "type": "string"
        },
        "speaker_Nmae": {
          "type": "string"
        }
      }
    }
  }
}